requirements.txt
wheel
setuptools
python-dotenv
black
pillow
django<5
django-debug-toolbar
django-imagekit
djangorestframework
djangorestframework-gis
django-cors-headers
djangorestframework-simplejwt
django-filter
psycopg2-binary